requirements.txt
asn1crypto==0.24.0
awscli==1.16.135
beautifulsoup4==4.6.0
botocore==1.12.125
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.7
configparser==3.5.0
crcmod==1.7
cryptography==2.1.4
decorator==4.3.0
dnspython==1.15.0
docutils==0.14
entrypoints==0.2.3.post1
enum34==1.1.6
future==0.17.1
futures==3.2.0
gdata==2.0.18
googlecl==0.9.13
html5lib==0.999999999
idna==2.6
ipaddress==1.0.17
jmespath==0.9.4
keyring==12.2.1
keyrings.alt==3.0
lxml==4.2.1
numpy==1.14.3
pathlib2==2.3.2
prettytable==0.7.2
psutil==5.4.2
pyasn1==0.4.5
pycrypto==2.6.1
pygobject==3.28.2
PyJWT==1.5.3
PyOpenGL==3.1.0
pyOpenSSL==17.5.0
pysmbc==1.0.15.6
python-dateutil==2.8.0
python-dotenv==0.10.3
python-telegram-bot==11.1.0
pyxdg==0.25
PyYAML==3.13
requests==2.18.4
rsa==3.4.2
s3transfer==0.2.0
scandir==1.7
scour==0.36
SecretStorage==2.3.1
six==1.11.0
typing==3.7.4
urllib3==1.22
webencodings==0.5
